requirements.txt
scikit-learn==1.7.0
pandas==2.3.0
numpy==1.26.4
PyYAML==6.0.1
streamlit==1.35.0
kagglehub 
ipykernel
pyyaml
-e .